requirements.txt
argon2-cffi==23.1.0

Django==4.2.4

django-cors-headers==4.2.0

django-environ==0.10.0

django-import-export==3.2.0

django-rest-framework==0.1.0

django-simple-history==3.3.0

djangorestframework==3.14.0

djangorestframework-simplejwt==5.2.2

djoser==2.2.0

openpyxl==3.1.2

PyJWT==2.8.0

whitenoise==6.5.0